% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundhm.R
\name{abundhm}
\alias{abundhm}
\title{Title}
\usage{
abundhm(tables, title)
}
\arguments{
\item{tables}{A table that contains differential metabolites}

\item{title}{The name of the pictures}
}
\value{
A double heatmap
}
\description{
Title
}
\examples{
#plot<-abundhm(tables,"CK & H")
#output<-"pic/doubleheatmap.png"
#ggsave(output, plot, width = 6, height = 6, dpi = 300)

}
