% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two number}
\usage{
add(x, y)
}
\arguments{
\item{x}{The explanation of the first param}

\item{y}{The explanation of the second param}
}
\value{
Get what through the function
}
\description{
A short description...
(if the line feed, there should be four space before a new line)
!!!Don't start with 'This package' or 'This function'
}
\details{
More detailed descriptions
}
\examples{
add(1,2)
\donttest{
add(3,4)}

}
