% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vip.R
\name{calc_vip}
\alias{calc_vip}
\title{Title Get the VIP value of a pls model}
\usage{
calc_vip(input, comgr)
}
\arguments{
\item{input}{the anaylzing filename}

\item{the}{compared groups}
}
\value{
a datframe of the vip value of all features
}
\description{
Title Get the VIP value of a pls model
}
\examples{
#input<- "data/metabolitesg.csv"
#comgr=c("CK","H")
#vip<-calc_vip(input,comgr)
}
