% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_met.R
\name{cor_met}
\alias{cor_met}
\title{Title Get the correlation picture}
\usage{
cor_met(input = "data/metabolitesg.csv")
}
\arguments{
\item{input}{the anaylzing filename}
}
\value{
a picture of the metabolites correlation map
}
\description{
Title Get the correlation picture
}
\examples{
#plot<-cor_met("data/metabolitesg.csv")
#plot
#ggsave(plot=plot,"pic/correlation_metabolites.png",width=20,height = 18,dpi=300,units="cm")
}
