% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_metmicro.R
\name{cor_metmicro}
\alias{cor_metmicro}
\title{Title Get the correlation map of Metabolites and microbiome}
\usage{
cor_metmicro(metabolites, microbiome)
}
\arguments{
\item{metabolites}{the metabolites table}

\item{microbiome}{the microbiome table}
}
\value{
a picture of the metabolites and microbiome correlation map
}
\description{
Title Get the correlation map of Metabolites and microbiome
}
\examples{
#p<-cor_metmicro(metabolites = "data/metabolitesg.csv", microbiome = "data/microbiome.csv")
#ggsave(plot=p,"pic/cor_met_micro.png",width=60,height = 20,dpi=300,units="cm")
}
