% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_micro.R
\name{cor_micro}
\alias{cor_micro}
\title{Title Get the correlation picture of microbiome}
\usage{
cor_micro(input = "data/microbiome.csv")
}
\arguments{
\item{input}{the anaylzing filename}
}
\value{
a picture of the microbiome correlation map
}
\description{
Title Get the correlation picture of microbiome
}
\examples{
#plot<-cor_met("data/microbiome.csv")
#plot
#ggsave(plot=plot,"pic/correlation_microbiome.png",width=20,height = 18,dpi=300,units="cm")
}
