% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornet.R
\name{cornet}
\alias{cornet}
\title{Title a function to make correlation network between genes and metabolites}
\usage{
cornet(
  gene,
  metabolites,
  geneedge,
  metabedge,
  genenode,
  metabnode,
  poscol,
  negcol
)
}
\arguments{
\item{gene}{a csv file containing gene abundance without group// example in data/gene.csv}

\item{metabolites}{a csv file containing peak concentration and group//example in data/metabolitesg.csv}

\item{geneedge}{gene edge color}

\item{metabedge}{metabolites edge color}

\item{genenode}{the color of genes nodes}

\item{metabnode}{the color of metabolites nodes}

\item{poscol}{if correlation no.>0, the color of the edge}

\item{negcol}{if correlation no.<0, the color of the edge}
}
\value{
the correlationnetwork
}
\description{
Title a function to make correlation network between genes and metabolites
}
\examples{
     setwd("D:/R_work/meTool")
     a<-cornet("data/gene.csv","data/metabolitesg.csv",geneedge="green",metabedge="pink",genenode="blue",metabnode="yellow",poscol="pink",negcol="green")

}
