% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSig.R
\name{getSig}
\alias{getSig}
\title{Title Obtaining significance through p-values.}
\usage{
getSig(p)
}
\arguments{
\item{p}{p value}
}
\value{
Significance shown in "\emph{\strong{","}","}"
}
\description{
Title Obtaining significance through p-values.
}
\examples{
getSig(0.04)
}
