% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelheatmap.R
\name{mantelheatmap}
\alias{mantelheatmap}
\title{Title A function to make heatmap with mantel test}
\usage{
mantelheatmap(metabolites, microbiome)
}
\arguments{
\item{metabolites}{A csv file contains sample names in the first column, Groups in the last column, and metabolites abundance in the middle.}

\item{micro}{Run micro<-microdiff("data/microbiome.csv") to find different microbiome.}

\item{num}{From which columns represent "Class", "Order", "Family", and "Genus"}
}
\value{
A picture
}
\description{
Title A function to make heatmap with mantel test
}
\examples{
#micro<-"data/microbiome.csv"
#metabolites<-"data/metabolitesg.csv"
#output<-"pic/micro_meta.png"
#p<-mantelheatmap(metabolites, micro)
#ggsave(output, p, width = 11, height = 6, dpi = 300)

}
