% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_pics.R
\name{met_pics}
\alias{met_pics}
\title{Title draw all the metabolites' abundance plot}
\usage{
met_pics(
  input = "data/metabolitesg.csv",
  G = c("CK", "L", "H"),
  color = c(CK = "violet", L = "lightblue2", H = "lightgreen"),
  filepath = "D:/R_work/meTool/pic/"
)
}
\arguments{
\item{input}{csv file//example: data/metabolitesg.csv}

\item{G}{the group need to be shown}

\item{color}{color of the group; the number of colors need to match the number of groups}

\item{filepath}{the output file path; all the pictures will be put in that file}
}
\value{
A batch of pictures
}
\description{
Title draw all the metabolites' abundance plot
}
\examples{
   #met_pics(input="data/metabolitesg.csv",G=c("CK","L","H"),color=c("CK"="violet","L"="lightblue2","H"="lightgreen"),filepath="D:/R_work/meTool/pic/")

}
