% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadiff.R
\name{metadiff}
\alias{metadiff}
\title{Title Find differential metabolites through t test}
\usage{
metadiff(input, comgr = c("CK", "H"))
}
\arguments{
\item{input}{Input file name and its file place}

\item{comgr}{The two groups you want to compare}
}
\value{
A table contains differential metabolites abundance and their P values
}
\description{
Title Find differential metabolites through t test
}
\examples{
#input<- "data/metabolite.csv"
#comgr=c("CK","H")
#tables<-metadiff(input,comgr)
}
