% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadiff_all.R
\name{metadiff_all}
\alias{metadiff_all}
\title{Title Find differential metabolites through t test, but export them all}
\usage{
metadiff_all(input, comgr = c("CK", "H"))
}
\arguments{
\item{input}{Input file name and its file place}

\item{comgr}{The two groups you want to compare}
}
\value{
A table contains differential metabolites abundance and their P values
}
\description{
Title Find differential metabolites through t test, but export them all
}
\examples{
#input<- "data/metabolitesg.csv"
#comgr=c("CK","H")
#tables<-metadiff_all(input,comgr)
}
