% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaflux.R
\name{metaflux}
\alias{metaflux}
\title{Title A function to cope with metabolic flux}
\usage{
metaflux(
  input,
  CKrepeat,
  EXPrepeat,
  EXPgroup = length(EXPrepeat),
  deltaRT = 0.01,
  ddC = 2e-04,
  Cnum = 6,
  deltaC = 1.0033
)
}
\arguments{
\item{input}{the data you want to put in;see data/SXT-neg.csv}

\item{CKrepeat}{the repeat count of Control group}

\item{EXPrepeat}{the repeat count of every Control group
NOTICE: it can be c(3,3) or 3. However, make sure every experiment group has the same repeat count}

\item{EXPgroup}{(Optional) the number of experiment groups( except CK ); it can count automatically}

\item{deltaRT}{the allowed deviation of retention time (default=0.01)}

\item{ddC}{the allowed deviation of the carbon atom (default =0.0002)}

\item{Cnum}{the number of carbon atoms that are marked by 13C (default =6)}

\item{deltaC}{the mass of a carbon atom(default=1.0033)}
}
\value{
a data frame that shows which peak may be the result of 13C isotope labelling
}
\description{
Sometimes it may be slow if the dataset is big. Just be patient.
all the data used are save in /data/ file
}
\examples{
\donttest{
setwd("D:/R_work/meTool")
neg25<-metaflux("data/20230728cell_flux_25pos.csv", CKrepeat = 3, EXPrepeat = 3, Cnum = 6)

#neg<-metaflux("data/SXT-neg.csv", CKrepeat = 3, EXPrepeat = c(3,3), Cnum = 6)}

}
