% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metmicro2dplot.R
\name{metmicro2dplot}
\alias{metmicro2dplot}
\title{Title draw pls-da plot}
\usage{
metmicro2dplot(
  metabolites,
  microbiome,
  sub = c("glutamate", "urea"),
  color = c(CK = "violet", L = "lightblue2", H = "lightgreen")
)
}
\arguments{
\item{metabolites}{metabolites csv file}

\item{microbiome}{microbiome csv file}

\item{sub}{the metabolites or microbiome name}

\item{color}{color of the group; the number of colors need to match the number of groups}
}
\value{
2d relative Abundance plot
}
\description{
Title draw pls-da plot
}
\examples{
#metabolites<-"data/metabolitesg.csv"
#microbiome<-"data/microbiome.csv"
#sub=c("glutamate","urea")
#color<-c("CK"="violet","L"="lightblue2","H"="lightgreen")
#p1<-metmicro2dplot(metabolites,microbiome,sub,color)
#ggsave("pic/glutamate_urea.png", plot = p1, width = 20, height = 20,dpi = 300,units = "cm")
}
