% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micmet_hm.R
\name{micmet}
\alias{micmet}
\title{Title The abundance heatmap of all microbiome and metabolites}
\usage{
micmet(microbiome, metabolites, comgr, output)
}
\arguments{
\item{microbiome}{the microbiome table}

\item{metabolites}{the metabolites table}

\item{comgr}{the compared groups(2 and only 2)}
}
\value{
a picture of all the metabolites and microbiome relative abundance
}
\description{
Title The abundance heatmap of all microbiome and metabolites
}
\examples{
#micmet(microbiome="data/microbiome.csv", metabolites="data/metabolitesg.csv",comgr=c("CK","L"),output="pic/circle_hp1.png")

}
