% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_heatmap.R
\name{micro_heatmap}
\alias{micro_heatmap}
\title{Title draw differential microbiome}
\usage{
micro_heatmap(input, comgr)
}
\arguments{
\item{input}{csv file//example: data/microbiome.csv}

\item{comgr}{compared group names(just 2)}
}
\value{
the microbiome abundance heatmap
}
\description{
Title draw differential microbiome
}
\examples{
#input<- "data/microbiome.csv"
#comgr=c("CK","L")
#p<-metabo_heatmap(input,comgr)
#ggsave(plot = p,"pic/metabolites_CK_L.png",width = 20,height = 15,dpi = 300,units = "cm")
}
