% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Title pathway plot}
\usage{
pathway(input)
}
\arguments{
\item{input}{pathway table}
}
\value{
pathway fig
}
\description{
Title pathway plot
}
\examples{
#input<-"data/pathway_CK_H.xlsx"
#p<-pathway(input)
#ggsave(plot = p,'pic/pathway_CK_H.png',width = 7.5, height = 6, dpi = 300)
}
