% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot.R
\name{pcaplot}
\alias{pcaplot}
\title{Title draw 2D pca plot}
\usage{
pcaplot(input, color = c(CK = "violet", L = "lightblue2", H = "lightgreen"))
}
\arguments{
\item{input}{csv file//example: data/metabolitesg.csv}

\item{color}{color of the group; the number of colors need to match the number of groups}
}
\value{
the PCA plot
}
\description{
Title draw 2D pca plot
}
\examples{
     setwd("D:/R_work/meTool")
     a<-pcaplot("data/metabolitesg.csv",color= c("CK"="violet","L"="lightblue2","H"="lightgreen"))
}
