% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarmass.R
\name{similarmass}
\alias{similarmass}
\title{Title find similar mass}
\usage{
similarmass(input1, input2, deltaRT = 0.05, deltamass = 1000)
}
\arguments{
\item{input1}{the filename of the first csv//  example: data/20230906_C18_5min_QC_MS1_NEG.csv}

\item{input2}{the filename of the second csv//  example: data/20230906_C18_5min_QC_MS1_POS.csv}

\item{deltaRT}{the deviation of RT; If not the same chromatographic column, deltaRT=NULL}

\item{deltamass}{the deviation of mass; unit:ppm}
}
\value{
a dataframe showing similar mass
}
\description{
Title find similar mass
}
\examples{
 setwd("D:/R_work/meTool")
# a<-similarmass("data/20230906_C18_5min_QC_MS1_NEG.csv","data/20230906_C18_5min_QC_MS1_POS.csv",deltaRT = 0.05,deltamass =500)
# b<-similarmass("data/20230906_C18_5min_QC_MS1_NEG.csv","data/20230906_HILIC_5min_QC_MS1_NEG.csv",deltaRT = NULL,deltamass =500)
# c<-similarmass("data/20230906_C18_5min_QC_MS1_POS.csv","data/20230906_HILIC_5min_QC_MS1_POS.csv",deltaRT = NULL,deltamass =500)
# d<-similarmass("data/20230906_HILIC_5min_QC_MS1_POS.csv","data/20230906_HILIC_5min_QC_MS1_NEG.csv",deltaRT = NULL,deltamass =500)
#write.csv(a,file = "output/C18_POS_NEG.csv")
#write.csv(b,file = "output/C18_HILIC_NEG.csv")
#write.csv(c,file = "output/C18_HILIC_NEG_POS.csv")
#write.csv(d,file = "output/HILIC_POS_NEG.csv")
}
