% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipp_plot.R
\name{vipp_plot}
\alias{vipp_plot}
\title{Title Get the VIP value of a pls model}
\usage{
vipp_plot(table, G = c("CK", "H"), color = c("#86FF86", "#E0FFE0"))
}
\arguments{
\item{input}{the anaylzing filename}

\item{the}{compared groups}
}
\value{
a datframe of the vip value of all features
}
\description{
Title Get the VIP value of a pls model
}
\examples{
#input<- "data/metabolitesg.csv"
#comgr=c("CK","H")
#tables<-metadiff(input,comgr)
#vipp_plot(tables,G=comgr)
}
