% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Title draw volcano plot}
\usage{
volcanoplot(
  input,
  CK = 1:6,
  TR = 13:18,
  legendp = c(0.73, 0.8),
  xlab = c(-2.5, 2.5),
  ylab = c(0, 4.6),
  textsize = 0.8
)
}
\arguments{
\item{input}{the metabolites file//example:data/metabolitesg.csv}

\item{CK}{the Control group rows}

\item{TR}{the treatment group rows}

\item{legendp}{the position of legend}

\item{xlab}{the range of abscissa}

\item{ylab}{the range of ordinate}

\item{textsize}{the text size of the point}
}
\value{
the volcano plot
}
\description{
Title draw volcano plot
}
\examples{
   setwd("D:/R_work/meTool")
   a<-volcanoplot("data/metabolitesg.csv",CK=1:6, TR=13:18,legendp=c(0.8,0.1), xlab=c(-2,2),ylab=c(0,5),textsize=0.8)
   a


}
