% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_heatmap.R
\name{metabo_heatmap}
\alias{metabo_heatmap}
\title{Title draw differential metabolites}
\usage{
metabo_heatmap(input, comgr, table)
}
\arguments{
\item{input}{csv file//example: data/metabolitesg.csv}

\item{comgr}{compared group names(just 2)}

\item{table}{table from function metadiff}
}
\value{
the metabolites abundance heatmap
}
\description{
Title draw differential metabolites
}
\examples{
#input<- "data/metabolitesg.csv"
#comgr=c("CK","L")
#table<-metadiff(input,comgr2)#筛选
#p<-metabo_heatmap(input,comgr,table)
#ggsave(plot = p,"pic/metabolites_CK_L.png",width = 20,height = 15,dpi = 300,units = "cm")
}
