% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdiff.R
\name{microdiff}
\alias{microdiff}
\title{Title Find differential bacteria}
\usage{
microdiff(dataname, comgr = c("CK", "L"))
}
\arguments{
\item{dataname}{The file name of microbiome abundance}

\item{comgr}{the compared groups}
}
\value{
Differential bacteria
}
\description{
Title Find differential bacteria
}
\examples{
#a<-microdiff("data/microbiome.csv",comgr=c("CK","L"))

}
