% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Title draw volcano plot}
\usage{
volcanoplot(input, titlename = "volcano plot", G = c("CK", "L"))
}
\arguments{
\item{input}{the metabolites file//example:data/metabolitesg.csv}

\item{titlename}{Title name}

\item{G}{compared groups}
}
\value{
the volcano plot
}
\description{
Title draw volcano plot
}
\examples{
   #setwd("D:/R_work/meTool")
   #G=c("CK","H")
   #a<-volcanoplot("data/metabolitesg.csv",G=G)
   #a


}
